requirements.txt
click==8.1.8
colorama==0.4.6
filelock==3.16.1
Flask==2.2.0
fsspec==2024.12.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.23.4
opencv-python==4.7.0.72
Pillow==9.2.0
sympy==1.13.1
torch==2.5.1
torchvision==0.20.1
typing_extensions==4.12.2
Werkzeug==2.2.0
